# swagger.yaml
swagger: "2.0"
info:
  description: "3MERN API NASA"
  version: "1.0.0"
  title: "NASA - 3MERN"
host: "petstore.swagger.io"
...
schemes:
- "http"
paths:
...
  /user:
    post:
      tags:
      - "user"
      summary: "Créer un utilisateur"
      description: "Création d'un utilisateur"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        "405":
          description: "Invalid input"
    get:
      tags:
      - "user"
      summary: "Récupérer les utilisateurs"
      description: "Récupérer un ou tous les utilisateurs"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "header"
        required: false
        type: "integer"
      responses:
        "200":
          description: "successful operation"
        "400":
          description: "Invalid status value"
    put:
      tags:
      - "user"
      summary: "Modifier un utilisateur"
      description: "Modifier d'un utilisateur"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        "405":
          description: "Invalid input"
    delete:
      tags:
      - "user"
      summary: "Supprimer un utilisateur"
      description: "Suppression d'un utilisateur"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "header"
        required: true
        type: "integer"
      responses:
        "400":
          description: "Invalid ID supplied"
...
definitions:
...
  User:
    type: "object"
    properties:
      email:
        type: "string"
      pseudo:
        type: "string"
      password:
        type: "string"
      isAdmin:
        type: "boolean"
        default: false
    xml:
      name: "User"
...